# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Dispatch API
  version: 3.0.0
  description: |
    Open-source bulk email campaign platform with tracking, multi-provider support, and batch processing.

    ## Authentication
    Most endpoints require authentication via a `session_token` cookie obtained from the login endpoint.
    Alternatively, an API key can be passed via the `X-API-Key` header.

    ## CSRF Protection
    All mutation endpoints (POST, PUT, DELETE) require a valid `X-CSRF-Token` header.
    The CSRF token is issued as a cookie on every response.

    ## Common Response Format
    All endpoints return a standard envelope:
    ```json
    {
      "success": true,
      "data": { ... },
      "message": "Human-readable message",
      "meta": { "pagination": { ... } }
    }
    ```
  contact:
    name: Dispatch
    url: https://github.com/iamjpsingh/bulk-email-sender
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_token
      description: Session token cookie set on login/register.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for programmatic access.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF token required for POST/PUT/DELETE requests. Value is read from the `csrf_token` cookie.

  # -------------------------------------------------------------------------
  # Reusable Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          description: Response payload (shape varies per endpoint).
        message:
          type: string
        meta:
          type: object
          properties:
            pagination:
              $ref: '#/components/schemas/Pagination'
      required: [success]
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Something went wrong
        error:
          type: string
      required: [success, message]
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5
        hasMore:
          type: boolean
          example: true

    # -- Auth --
...
    EmailLog:
      type: object
      properties:
        id:
          type: string
        recipient_email:
          type: string
          format: email
        recipient_name:
          type: string
        subject:
          type: string
        status:
          type: string
        send_type:
          type: string
        provider_type:
          type: string
        config_name:
          type: string
        sent_at:
          type: string
          format: date-time
        opened_at:
          type: string
          format: date-time
          nullable: true
        open_count:
          type: integer
        click_count:
          type: integer
    LogStats:
      type: object
      properties:
        sent:
          type: integer
        failed:
          type: integer
        total:
          type: integer

    # -- Dashboard --
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
...
    SearchParam:
      name: search
      in: query
      schema:
        type: string

# ---------------------------------------------------------------------------
# Default Security
# ---------------------------------------------------------------------------
security:
  - cookieAuth: []
  - apiKeyAuth: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /report/logs:
    get:
      tags: [Reports]
      summary: Get email logs with filtering
      operationId: getEmailLogs
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: send_type
          in: query
          schema:
            type: string
        - name: provider
          in: query
          schema:
            type: string
        - name: campaign_id
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/SearchParam'
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated email logs with stats.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          logs:
                            type: array
                            items:
                              $ref: '#/components/schemas/EmailLog'
                          stats:
                            $ref: '#/components/schemas/LogStats'
                          pagination:
                            $ref: '#/components/schemas/Pagination'
...
  /report/export/csv:
    get:
      tags: [Reports]
      summary: Export logs as CSV
      operationId: exportLogsCsv
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: send_type
          in: query
          schema:
            type: string
        - name: provider
          in: query
          schema:
            type: string
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: CSV file download.
          content:
            text/csv:
              schema:
                type: string
  /report/export/json:
    get:
      tags: [Reports]
      summary: Export logs as JSON
      operationId: exportLogsJson
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: send_type
          in: query
          schema:
            type: string
        - name: provider
          in: query
          schema:
            type: string
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: JSON file download.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmailLog'
  /report/logs/{id}:
    delete:
      tags: [Reports]
      summary: Delete a single email log
      operationId: deleteLog
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Log deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /report/logs/delete-bulk:
    post:
      tags: [Reports]
      summary: Bulk delete email logs
      operationId: bulkDeleteLogs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items:
                    type: string
              required: [ids]
      responses:
        '200':
          description: Logs deleted.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          deleted:
                            type: integer
        '400':
          description: No log IDs provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # =========================================================================
  # Dashboard
  # =========================================================================
...